+++ src/main/java/com/nguyenquyhy/discordbridge/listeners/DeathListener.java
@@ -5,8 +5,8 @@ import com.nguyenquyhy.discordbridge.models.ChannelConfig;
 import com.nguyenquyhy.discordbridge.models.GlobalConfig;
 import com.nguyenquyhy.discordbridge.utils.ChannelUtil;
 import com.nguyenquyhy.discordbridge.utils.ConfigUtil;
-import de.btobastian.javacord.DiscordAPI;
-import de.btobastian.javacord.entities.Channel;
+import net.dv8tion.jda.core.JDA;
+import net.dv8tion.jda.core.entities.TextChannel;
 import org.apache.commons.lang3.StringUtils;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.event.Listener;
@@ -17,7 +17,7 @@ public class DeathListener {
     public void onPlayerDeath(DestructEntityEvent.Death event) {
         DiscordBridge mod = DiscordBridge.getInstance();
         GlobalConfig config = mod.getConfig();
-        DiscordAPI client = mod.getBotClient();
+        JDA client = mod.getBotClient();
 
         if (!(event.getTargetEntity() instanceof Player) || event.isMessageCancelled() || StringUtils.isBlank(event.getMessage().toPlain())) return;
         Player player = (Player) event.getTargetEntity();
@@ -27,7 +27,7 @@ public class DeathListener {
                 if (StringUtils.isNotBlank(channelConfig.discordId) && channelConfig.discord != null) {
                     String template = ConfigUtil.get(channelConfig.discord.deathTemplate, null);
                     if (StringUtils.isNotBlank(template)) {
-                        Channel channel = client.getChannelById(channelConfig.discordId);
+                        TextChannel channel = client.getTextChannelById(channelConfig.discordId);
                         ChannelUtil.sendMessage(channel, template.replace("%s", event.getMessage().toPlain()));
                     }
                 }