+++ src/main/java/com/nguyenquyhy/discordbridge/utils/DiscordUtil.java
@@ -1,19 +1,7 @@
 package com.nguyenquyhy.discordbridge.utils;
 
-import com.google.common.collect.Lists;
-import com.nguyenquyhy.discordbridge.models.ChannelMinecraftMentionConfig;
-import de.btobastian.javacord.entities.Channel;
-import de.btobastian.javacord.entities.Server;
-import de.btobastian.javacord.entities.User;
-import de.btobastian.javacord.entities.permissions.Role;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.http.util.TextUtils;
-import org.spongepowered.api.text.Text;
-import org.spongepowered.api.text.serializer.TextSerializers;
+import net.dv8tion.jda.core.entities.*;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 public class DiscordUtil {
@@ -23,10 +11,10 @@ public class DiscordUtil {
      * @param server The server to search through Users
      * @return The User, if any, that matches the name supplied
      */
-    static Optional<User> getUserByName(String name, Server server) {
-        for (User user : server.getMembers()) {
-            if (user.getName().equalsIgnoreCase(name) || (user.getNickname(server) != null && user.getNickname(server).equalsIgnoreCase(name))) {
-                return Optional.of(user);
+    static Optional<Member> getMemberByName(String name, Guild server) {
+        for (Member member : server.getMembers()) {
+            if (member.getEffectiveName().equalsIgnoreCase(name) || (member.getNickname() != null && member.getNickname().equalsIgnoreCase(name))) {
+                return Optional.of(member);
             }
         }
         return Optional.empty();
@@ -37,7 +25,7 @@ public class DiscordUtil {
      * @param server The server to search through Roles
      * @return The Role, if any, that matches the name supplied
      */
-    static Optional<Role> getRoleByName(String name, Server server) {
+    static Optional<Role> getRoleByName(String name, Guild server) {
         for (Role role : server.getRoles()) {
             if (role.getName().equalsIgnoreCase(name)) {
                 return Optional.of(role);
@@ -51,8 +39,8 @@ public class DiscordUtil {
      * @param server The server to search through Roles
      * @return The Channel, if any, that matches the name supplied
      */
-    static Optional<Channel> getChannelByName(String name, Server server) {
-        for (Channel channel : server.getChannels()) {
+    static Optional<Channel> getChannelByName(String name, Guild server) {
+        for (Channel channel : server.getTextChannels()) {
             if (channel.getName().equalsIgnoreCase(name)) {
                 return Optional.of(channel);
             }